The IByteWriter interface has been added, allowing you to write bytes to an array. A ByteWriter class has been added that implements IByteWriter. Part of the compiler has also been rewritten to use IByteWriter.

+++ PettyLang/Semantic/Scope.cs
@@ -22,9 +22,9 @@ public class Scope
     private Dictionary<string, FunctionSymbol> functions = new();
     private Dictionary<string, ClassSymbol> classes = new();
     
-    private static ulong globalFreeID = 0;
+    private static int globalFreeID = 0;
 
-    public ulong GetFreeID()
+    public int GetFreeID()
     {
         return globalFreeID++;
     }
